Improve Gmail transport reuse and registry structure
