+++ Sources/Mailozaurr.Application/MailMessageActionPlanRegistryService.cs
@@ -3,7 +3,7 @@ namespace Mailozaurr.Application;
 /// <summary>
 /// Implements reusable lifecycle operations for persisted message action plan batches.
 /// </summary>
-public sealed class MailMessageActionPlanRegistryService : IMailMessageActionPlanRegistryService {
+public sealed partial class MailMessageActionPlanRegistryService : IMailMessageActionPlanRegistryService {
     private readonly IMailMessageActionPlanBatchStore _store;
     private readonly IMailMessageActionPlanExchangeService _exchangeService;
     private readonly IMailMessageActionPreviewService _previewService;
@@ -448,343 +448,4 @@ public sealed class MailMessageActionPlanRegistryService : IMailMessageActionPla
         return OperationResult.Success();
     }
 
-    private static MessageActionExecutionPlan ClonePlan(MessageActionExecutionPlan plan) => new() {
-        Succeeded = plan.Succeeded,
-        Code = plan.Code,
-        Message = plan.Message,
-        Name = plan.Name,
-        Summary = plan.Summary,
-        Action = plan.Action,
-        ExecutionKind = plan.ExecutionKind,
-        ProfileId = plan.ProfileId,
-        MailboxId = plan.MailboxId,
-        FolderId = plan.FolderId,
-        RequestedCount = plan.RequestedCount,
-        UniqueMessageCount = plan.UniqueMessageCount,
-        MessageIds = plan.MessageIds.ToList(),
-        RequestedDestinationFolderId = plan.RequestedDestinationFolderId,
-        Destination = plan.Destination == null
-            ? null
-            : new MailFolderTargetResolution {
-                ProfileId = plan.Destination.ProfileId,
-                MailboxId = plan.Destination.MailboxId,
-                RequestedValue = plan.Destination.RequestedValue,
-                IsAlias = plan.Destination.IsAlias,
-                Alias = plan.Destination.Alias,
-                IsSupported = plan.Destination.IsSupported,
-                IsResolved = plan.Destination.IsResolved,
-                EffectiveFolderId = plan.Destination.EffectiveFolderId,
-                FolderDisplayName = plan.Destination.FolderDisplayName,
-                FolderPath = plan.Destination.FolderPath,
-                Summary = plan.Destination.Summary
-            },
-        DesiredState = plan.DesiredState,
-        ConfirmationToken = plan.ConfirmationToken,
-        ConfirmationProvided = plan.ConfirmationProvided,
-        ConfirmationValidated = plan.ConfirmationValidated,
-        Warnings = plan.Warnings.ToList()
-    };
-
-    private static MessageActionExecutionPlan TransformPlan(MessageActionExecutionPlan plan, MessageActionPlanBatchTransformRequest transform) {
-        var cloned = ClonePlan(plan);
-        var transformedProfileId = string.IsNullOrWhiteSpace(transform.ProfileId) ? cloned.ProfileId : transform.ProfileId!.Trim();
-        var transformedMailboxId = string.IsNullOrWhiteSpace(transform.MailboxId) ? cloned.MailboxId : transform.MailboxId!.Trim();
-        var transformedFolderId = string.IsNullOrWhiteSpace(transform.FolderId) ? cloned.FolderId : transform.FolderId!.Trim();
-
-        cloned.ProfileId = transformedProfileId;
-        cloned.MailboxId = transformedMailboxId;
-        cloned.FolderId = transformedFolderId;
-
-        if (cloned.Destination != null) {
-            cloned.Destination.ProfileId = transformedProfileId;
-            cloned.Destination.MailboxId = transformedMailboxId;
-        }
-
-        if (!string.IsNullOrWhiteSpace(transform.DestinationFolderId) && string.Equals(cloned.ExecutionKind, "Move", StringComparison.OrdinalIgnoreCase)) {
-            var transformedDestination = transform.DestinationFolderId!.Trim();
-            cloned.RequestedDestinationFolderId = transformedDestination;
-            cloned.Destination = null;
-        }
-
-        cloned.ConfirmationProvided = false;
-        cloned.ConfirmationValidated = false;
-        cloned.ConfirmationToken = CreateConfirmationToken(cloned);
-        cloned.Summary = BuildStoredPlanSummary(cloned);
-        return cloned;
-    }
-
-    private static IReadOnlyList<MessageActionPreviewItem> SelectCommonActions(CommonMessageActionsPreview preview, IReadOnlyList<string>? actions) {
-        if (actions == null || actions.Count == 0) {
-            return preview.Actions
-                .Where(action => action.Succeeded)
-                .ToArray();
-        }
-
-        var selected = new List<MessageActionPreviewItem>();
-        foreach (var actionName in actions
-                     .Where(action => !string.IsNullOrWhiteSpace(action))
-                     .Select(action => action.Trim())
-                     .Distinct(StringComparer.OrdinalIgnoreCase)) {
-            var match = preview.Actions.FirstOrDefault(action =>
-                string.Equals(action.Action, actionName, StringComparison.OrdinalIgnoreCase) &&
-                action.Succeeded);
-            if (match != null) {
-                selected.Add(match);
-            }
-        }
-
-        return selected;
-    }
-
-    private static string? ResolveDestinationFolderId(MessageActionPreviewItem action, string? fallbackDestinationFolderId) =>
-        string.Equals(action.Action, "move", StringComparison.OrdinalIgnoreCase)
-            ? action.Destination?.EffectiveFolderId ?? action.RequestedDestinationFolderId ?? fallbackDestinationFolderId
-            : null;
-
-    private static string? CreateConfirmationToken(MessageActionExecutionPlan plan) =>
-        plan.ExecutionKind switch {
-            "Move" => !string.IsNullOrWhiteSpace(plan.Destination?.EffectiveFolderId ?? plan.RequestedDestinationFolderId)
-                ? MessageActionConfirmationTokens.CreateMoveToken(
-                    plan.ProfileId,
-                    plan.MailboxId,
-                    plan.FolderId,
-                    plan.MessageIds,
-                    plan.Destination?.EffectiveFolderId ?? plan.RequestedDestinationFolderId!)
-                : null,
-            "Delete" => MessageActionConfirmationTokens.CreateDeleteToken(
-                plan.ProfileId,
-                plan.MailboxId,
-                plan.FolderId,
-                plan.MessageIds),
-            "SetReadState" when plan.DesiredState.HasValue => MessageActionConfirmationTokens.CreateReadStateToken(
-                plan.ProfileId,
-                plan.MailboxId,
-                plan.FolderId,
-                plan.MessageIds,
-                plan.DesiredState.Value),
-            "SetFlaggedState" when plan.DesiredState.HasValue => MessageActionConfirmationTokens.CreateFlaggedStateToken(
-                plan.ProfileId,
-                plan.MailboxId,
-                plan.FolderId,
-                plan.MessageIds,
-                plan.DesiredState.Value),
-            _ => plan.ConfirmationToken
-        };
-
-    private static string BuildTransformSummary(MessageActionExecutionPlan sourcePlan, MessageActionExecutionPlan transformedPlan, bool willChange, bool tokenChanged) {
-        var changes = new List<string>();
-        if (!string.Equals(sourcePlan.ProfileId, transformedPlan.ProfileId, StringComparison.Ordinal)) {
-            changes.Add($"profile {sourcePlan.ProfileId}->{transformedPlan.ProfileId}");
-        }
-        if (!string.Equals(sourcePlan.MailboxId, transformedPlan.MailboxId, StringComparison.Ordinal)) {
-            changes.Add($"mailbox {(sourcePlan.MailboxId ?? "<none>")}->{(transformedPlan.MailboxId ?? "<none>")}");
-        }
-        if (!string.Equals(sourcePlan.FolderId, transformedPlan.FolderId, StringComparison.Ordinal)) {
-            changes.Add($"folder {(sourcePlan.FolderId ?? "<none>")}->{(transformedPlan.FolderId ?? "<none>")}");
-        }
-        if (!string.Equals(sourcePlan.RequestedDestinationFolderId, transformedPlan.RequestedDestinationFolderId, StringComparison.Ordinal)) {
-            changes.Add($"destination {(sourcePlan.RequestedDestinationFolderId ?? "<none>")}->{(transformedPlan.RequestedDestinationFolderId ?? "<none>")}");
-        }
-        if (tokenChanged) {
-            changes.Add("confirmation token regenerated");
-        }
-
-        return willChange || tokenChanged
-            ? $"{sourcePlan.Action}: {string.Join(", ", changes)}"
-            : $"{sourcePlan.Action}: unchanged";
-    }
-
-    private static IReadOnlyList<SelectedPlan> SelectPlans(
-        MailMessageActionPlanBatch source,
-        MessageActionPlanBatchTransformRequest transform,
-        out OperationResult? error) {
-        error = null;
-
-        var selected = new Dictionary<int, SelectedPlan>();
-        var hasIndexes = transform.PlanIndexes != null && transform.PlanIndexes.Count > 0;
-        var hasNames = transform.PlanNames != null && transform.PlanNames.Count > 0;
-        IEnumerable<int> requestedIndexes = hasIndexes ? transform.PlanIndexes! : Array.Empty<int>();
-        IEnumerable<string> requestedNames = hasNames ? transform.PlanNames! : Array.Empty<string>();
-
-        if (!hasIndexes && !hasNames) {
-            return source.Plans
-                .Select((plan, index) => new SelectedPlan(index, plan))
-                .ToArray();
-        }
-
-        if (hasIndexes) {
-            foreach (var index in requestedIndexes.Distinct()) {
-                if (index < 0 || index >= source.Plans.Count) {
-                    error = OperationResult.Failure("action_plan_batch_index_invalid", $"Action plan batch '{source.Id}' does not contain a plan at index {index}.");
-                    return Array.Empty<SelectedPlan>();
-                }
-
-                selected[index] = new SelectedPlan(index, source.Plans[index]);
-            }
-        }
-
-        if (hasNames) {
-            foreach (var requestedName in requestedNames
-                         .Where(name => !string.IsNullOrWhiteSpace(name))
-                         .Select(name => name.Trim())
-                         .Distinct(StringComparer.OrdinalIgnoreCase)) {
-                var matches = source.Plans
-                    .Select((plan, index) => new SelectedPlan(index, plan))
-                    .Where(selection => PlanMatchesName(selection.Plan, requestedName))
-                    .ToArray();
-                if (matches.Length == 0) {
-                    error = OperationResult.Failure("action_plan_batch_name_invalid", $"Action plan batch '{source.Id}' does not contain a plan named '{requestedName}'.");
-                    return Array.Empty<SelectedPlan>();
-                }
-
-                foreach (var match in matches) {
-                    selected[match.Index] = match;
-                }
-            }
-        }
-
-        if (selected.Count == 0) {
-            error = OperationResult.Failure("action_plan_batch_invalid", "Action plan batch transform must include at least one plan.");
-            return Array.Empty<SelectedPlan>();
-        }
-
-        return selected
-            .OrderBy(item => item.Key)
-            .Select(item => item.Value)
-            .ToArray();
-    }
-
-    private static MailMessageActionPlanBatchCompact ToCompact(MailMessageActionPlanBatch batch) {
-        var profileCount = batch.Plans
-            .Select(plan => plan.ProfileId)
-            .Where(profileId => !string.IsNullOrWhiteSpace(profileId))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .Count();
-        var readyPlanCount = batch.Plans.Count(plan => plan.Succeeded);
-
-        return new MailMessageActionPlanBatchCompact {
-            Id = batch.Id,
-            Name = batch.Name,
-            PlanCount = batch.Plans.Count,
-            ReadyPlanCount = readyPlanCount,
-            ProfileCount = profileCount,
-            PlanNames = batch.Plans
-                .Select(plan => !string.IsNullOrWhiteSpace(plan.Name) ? plan.Name : BuildStoredPlanSummary(plan))
-                .Distinct(StringComparer.OrdinalIgnoreCase)
-                .ToList(),
-            UpdatedAt = batch.UpdatedAt,
-            Summary = $"{batch.Id} ({batch.Plans.Count} plan(s), {readyPlanCount} ready)"
-        };
-    }
-
-    private static MailMessageActionPlanBatchSummary ToSummary(MailMessageActionPlanBatch batch) {
-        var profileIds = batch.Plans
-            .Select(plan => plan.ProfileId)
-            .Where(profileId => !string.IsNullOrWhiteSpace(profileId))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .OrderBy(profileId => profileId, StringComparer.OrdinalIgnoreCase)
-            .ToList();
-        var readyPlanCount = batch.Plans.Count(plan => plan.Succeeded);
-        var actionCounts = batch.Plans
-            .Where(plan => !string.IsNullOrWhiteSpace(plan.Action))
-            .GroupBy(plan => plan.Action, StringComparer.OrdinalIgnoreCase)
-            .OrderBy(group => group.Key, StringComparer.OrdinalIgnoreCase)
-            .ToDictionary(group => group.Key, group => group.Count(), StringComparer.OrdinalIgnoreCase);
-
-        return new MailMessageActionPlanBatchSummary {
-            Id = batch.Id,
-            Name = batch.Name,
-            PlanCount = batch.Plans.Count,
-            ReadyPlanCount = readyPlanCount,
-            ProfileIds = profileIds,
-            ActionCounts = actionCounts,
-            PlanNames = batch.Plans
-                .Select(plan => !string.IsNullOrWhiteSpace(plan.Name) ? plan.Name : BuildStoredPlanSummary(plan))
-                .Distinct(StringComparer.OrdinalIgnoreCase)
-                .ToList(),
-            UpdatedAt = batch.UpdatedAt,
-            Summary = $"{batch.Id} ({batch.Plans.Count} plan(s), {readyPlanCount} ready, {actionCounts.Count} action type(s))"
-        };
-    }
-
-    private static IReadOnlyList<MailMessageActionPlanBatch> ApplyBatchQuery(
-        IReadOnlyList<MailMessageActionPlanBatch> batches,
-        MailMessageActionPlanBatchQuery? query) {
-        if (query == null) {
-            return batches;
-        }
-
-        var requestedPlanNames = query.PlanNames
-            .Where(name => !string.IsNullOrWhiteSpace(name))
-            .Select(name => name.Trim())
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-        var requestedProfileIds = query.ProfileIds
-            .Where(profileId => !string.IsNullOrWhiteSpace(profileId))
-            .Select(profileId => profileId.Trim())
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-        var requestedActions = query.Actions
-            .Where(action => !string.IsNullOrWhiteSpace(action))
-            .Select(action => action.Trim())
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        IEnumerable<MailMessageActionPlanBatch> filtered = batches;
-        if (requestedPlanNames.Length > 0 || requestedProfileIds.Length > 0 || requestedActions.Length > 0) {
-            filtered = filtered.Where(batch =>
-                (requestedPlanNames.Length == 0 || batch.Plans.Any(plan => requestedPlanNames.Any(requestedName => PlanMatchesName(plan, requestedName)))) &&
-                (requestedProfileIds.Length == 0 || batch.Plans.Any(plan => requestedProfileIds.Any(requestedProfileId => string.Equals(plan.ProfileId, requestedProfileId, StringComparison.OrdinalIgnoreCase)))) &&
-                (requestedActions.Length == 0 || batch.Plans.Any(plan => requestedActions.Any(requestedAction => string.Equals(plan.Action, requestedAction, StringComparison.OrdinalIgnoreCase)))));
-        }
-
-        var ordered = ApplyBatchQuerySort(filtered, query.SortBy, query.Descending);
-        return ordered.ToArray();
-    }
-
-    private static IEnumerable<MailMessageActionPlanBatch> ApplyBatchQuerySort(
-        IEnumerable<MailMessageActionPlanBatch> batches,
-        MailMessageActionPlanBatchSortBy sortBy,
-        bool descending) {
-        Func<MailMessageActionPlanBatch, object> keySelector = sortBy switch {
-            MailMessageActionPlanBatchSortBy.Name => batch => batch.Name,
-            MailMessageActionPlanBatchSortBy.PlanCount => batch => batch.Plans.Count,
-            MailMessageActionPlanBatchSortBy.ReadyPlanCount => batch => batch.Plans.Count(plan => plan.Succeeded),
-            MailMessageActionPlanBatchSortBy.UpdatedAt => batch => batch.UpdatedAt,
-            MailMessageActionPlanBatchSortBy.ActionTypeCount => batch => batch.Plans
-                .Select(plan => plan.Action)
-                .Where(action => !string.IsNullOrWhiteSpace(action))
-                .Distinct(StringComparer.OrdinalIgnoreCase)
-                .Count(),
-            _ => batch => batch.Id
-        };
-
-        var ordered = descending
-            ? batches.OrderByDescending(keySelector)
-            : batches.OrderBy(keySelector);
-
-        return ordered.ThenBy(batch => batch.Id, StringComparer.OrdinalIgnoreCase);
-    }
-
-    private static bool PlanMatchesName(MessageActionExecutionPlan plan, string requestedName) =>
-        (!string.IsNullOrWhiteSpace(plan.Name) && string.Equals(plan.Name, requestedName, StringComparison.OrdinalIgnoreCase)) ||
-        (!string.IsNullOrWhiteSpace(plan.Summary) && string.Equals(plan.Summary, requestedName, StringComparison.OrdinalIgnoreCase));
-
-    private static string BuildStoredPlanSummary(MessageActionExecutionPlan plan) {
-        var messageCount = plan.UniqueMessageCount == 1 ? "1 message" : $"{plan.UniqueMessageCount} messages";
-        return !string.IsNullOrWhiteSpace(plan.Name)
-            ? $"{plan.Name} ({messageCount})"
-            : $"{plan.Action} ({messageCount})";
-    }
-
-    private sealed class SelectedPlan {
-        public SelectedPlan(int index, MessageActionExecutionPlan plan) {
-            Index = index;
-            Plan = plan;
-        }
-
-        public int Index { get; }
-
-        public MessageActionExecutionPlan Plan { get; }
-    }
 }
